langchain-models/requirements.txt
# Langchain Core
Langchain
langchain-core

# Open AI Integration
langchain-openai
openai

# Anthropic Integration
langchain-Anthropic

# Google Gemini (Palm) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn



